Initial text STL parser
A STL text file can be parsed and turned into a data structure
containing all of the floating point values required to define
a solid.  The primary limitation is that the parser cannot handle
named solids.  Committing the first working implementation.

This implementation is significant because it drops usage of
Flex and Bison as a parser generator and relies on Boost Spirit
instead.  This allows for native C++ interop, which is important
for processing text STL files because unlike binary STL files, it
is not known at the start of parsing how many facets will need to
be parsed.