Allow solids with names
Solids may be parsed with a name.  The name is not relevant data and
is discarded.  Equality between the first and last name is not verified
because it is not important when parsing floating point values from a
text STL file.

Applying a consistent coding style with this commit to be used going
forward.  Namespaces will never be imported with a using declarative
to avoid naming conflicts.  Classes, functions, structs, enums,
variables, etc. will be named according to Java conventions to better
differentiate project code from STL code.