Make meson build file functional
On Arch Linux, the antlr4-runtime package is missing a pkg-config file.
Supplying the missing pkg-config file allows usage of the dependency
function in meson, which tries to use pkg-config by default.

Implement more meson.build files to work with the desired directory
structure of placing all source files in the src folder.

Create and compile the generated files from ANTLR and link them together
with the final executable.  Do this to allow testing that the build
process is working.