Add command line parsing with CLI11
While Boost Program Options is a venerable and common implementation for
command line parsing with C++, there is quite a bit of boilerplate code
required for basic functionality.  Additionally, the output when using
the help option does not look like most other Unix-based programs.

Prefer to use command line parsing that is closer to argparse in python
and which has more of a natural interface.  When approaching the code
for the first time, it is straight-forward to understand at a
high-level.