Switch language to C#
Want to work in a language that has familiar syntax, a standard package
manager, excellent tooling support for a pleasurable development
experience, is fun to use, has a garbage collector, and isn't commonly
viewed as a beast due to the power, scope, and complexity of the
language.

While this makes it difficult to do graphical programming because the
GTK+ and OpenGL bindings are lacking for .NET Core, can start with a CLI
implementation and rely on external third-party graphical viewers.

Signed-off-by: George Kramer <george.kramer@live.com>