Add STL binary reader
Prefer not to use ANTLR for creating a STL binary reader.  In order to
get a floating point number, one must first match four bytes with the
lexer and parser infrastructure.  Then for each byte, run .GetText() on
the context of a byte ITerminalNode, get the first character of the
string as a char, and convert the char to a byte.  Then with an array of
bytes, convert to a floating point number.  This is quite a bit of extra
work for a simple file format, and there are a number of extra
conversions occurring with the fundamental data.  The undoubtedly adds
overhead beyond what a BinaryReader can offer.  While there may be a
simpler way, the documentation on reading binary files with ANTLR
doesn't seem to suggest this.  By-and-large, it also appears that ANTLR
is great for recognizing text streams and not binary streams.

Prefer to read the file as a stream and build up the facet array
incrementally.

Signed-off-by: George Kramer <george.kramer@live.com>