Make STL ASCII grammar more accurate
Allow floating point numbers with exponents to be tokenized.  Now, all
floating point number representations can be tokenized.

Remove tokens that do not need to be accessed programmatically and are
merely serving as anchors. There are no values of interest in these
tokens.  They only serve to structure data of interest.

Signed-off-by: George Kramer <george.kramer@live.com>