Add option for reading STL ASCII file
Automatically determining whether the format of a STL file is ASCII or
binary first requires an attempt at parsing.  This is a little
cumbersome and results in some slightly messy code.  Prefer to have the
user define the input format correctly.

Default to working with STL binary files because they are more compact.
If STL ASCII files are required, then the --ascii option may be used.
This option influences STL file read and write behavior.

Signed-off-by: George Kramer <george.kramer@live.com>