Add a 3D point tree for creating an index of edges
Edges can be identified uniquely by their midpoint for well-formed STL
files.  Implement a k-d tree that can natively use a Point3D key to
return a generic reference-type value.  A k-d tree guarantees proper
indexing when compared to hashing with a dictionary.  Require that all
edge-identifying midpoints be supplied up front to construct a balanced
tree.  Use in-place sorting with arrays to save memory and keep the
construction of the tree simple.

Implement an extension method for easily calculating the midpoint of an
edge on the fly.

Signed-off-by: George Kramer <george.kramer@live.com>