Remove GtkGLArea widget usage
While the GtkGLArea widget is the correct widget to use for doing work
with OpenGL in a GTK+ application, there are few examples on how to use
this properly and none are easily understood.  Most examples are in C
and even fewer examples exist in C#.

While OpenTK has better documentation and provides a chance to learn the
details of OpenGL programming, OpenTK 4.0 no longer includes support for
GLWidget.  GLWidget provided an interface between OpenTK and GtkWindow.
Usage always seemed to be done programmatically though, which conflicts
with the goal of the project to create the GUI using markup from Glade.
Given this is the future of OpenTK, it makes little sense to rely on
deprecated/non-existent methods.

The idea going forward is to experiment with an idea proposed by an
article that creates a multi-windowed application.  One window would be
the GtkWindow that contains the controls for inputting data.  The other
window would be an OpenTK GameWindow that receives the data via message
passing.  Each window would be able to operate in it's own native
context and can be developed using documentation that is readily
available.  The trick will be to get the message passing right so that
event triggers fire properly, along with the threading to allow both
windows to run well simultaneously.

Signed-off-by: George Kramer <george.kramer@live.com>