Add function for removing small unsupported regions
There are often cases where a region meets the requirements for being
supported, however, in reality, it's actually too small to really need
support.  Remove these regions from the list that is used to create
scaffolding.

The basic idea is to find the bounding vertices of the region, create a
list of large diagonals, and then find the angle between the large
diagonals.  If at any point the angle between two large diagonals is
nearly perpendicular, then the region is considered large enough to need
support.