Do not create wrapper function around simple LINQ
These are easy enough to read on their own and a wrapper function does
not provide much value.  This stands out in comparison to the boolean
function wrapper that determines whether a facet needs to be supported
because it names potentially complex logic that could be hard to
understand.  It is also a critical part of the program.