Remove stateful visited property from facet
The facet object was used to store visited state information.  This
state was only relevant to the execution of the process that generates
scaffolding, rather than being an inherent property of a facet.

To get around this, need to create a new facet index association.  Index
facets by their centroid.  Then associate a boolean variable, whether a
facet has been visited, with the facet.  Use this facet visited index to
support the process of growing an unsupported region.

State is now maintained by the creation and destruction of a data
structure as needed by the process, rather than manually in code by
resetting visited values.