Start the process of generating line scaffolding
The general idea is to start with a region and its center point.  A
support is needed in the center of the region.  Then on both sides of
the center support, iterate outward in the direction of the support
vector creating new supports as needed.  Model supports as planes to
generate intersection points with regions, which are a set of facets,
which are composed of lines to ultimately generate the intersection
between a line and a plane.