Finish initial implementation of generating line scaffolding
With the intersection points between lines and a plane, a point grid may
be created.  From there, every square in the point grid may be converted
into two triangles to produce a naive tessellation.

This required creation a Polygon3D object to keep track of a mesh, which
may be useful later to do additional processing on the mesh.  This
change allowed a Region and a Model to use the same base class, since
both are 3D polygons.