Unify definitions of point equality and comparison checks
The distinction between a Point3DComperer and a Point3DXComparer was not
obvious and added to the confusion of their use.  Make these so they may
be used in either or both contexts of equality and comparison checks.

As a result, when using these, define them according to their use based
on the declared type.