Simplify which planes are used to generate scaffolding
Prefer to find an optimistic amount of candidate planes based on a
simple calculation, and let the criteria for whether intersection points
are found with each plane handle whether they were valid, since it is
the ultimate determining factor.  Doing a bounding box check for whether
the point used to create the plane lies inside is not accurate.

Use 2D diagonal distance to estimate the size of the region, since the
planes are created in the XY direction only.  The diagonal represents
the largest possible size for both directions, which is why it may be
used in both cases.