Improve tessellation of triangles from point grid
Intersection points need to be properly sorted so they may be arranged
in increasing order along the line support.

Calculate an average segment count based on the Z position of each
intersection point and how long the segment length is.  This lets the
density of the scaffolding vary based on the density of the
intersections.