Switch to custom math library
Want to use floats rather than doubles to reduce the memory used by a STL
file in half.  This is also a more appropriate data type because it
avoids a translation in data types when writing a STL binary file and
when rendering with OpenGL.

Having full control over the math library allows for better integration
throughout the program.  Copy operations may be avoided when reading a
STL ASCII file that previously conformed to a rigid interface.  Vertices
may be moved a certain distance and the distance between them may be
calculated in both 2D and 3D.

Model how operators were overloaded with MathNet.Spatial.Euclidean since
this were very handy means of expression.