Generate contour scaffolding
Prefer a keep-it-simple initial approach by considering contour
scaffolding as groups of line supports defined by each bounding edge of
an unsupported regions.  This allows for more code reuse.

The adjacency information is correct still even though unsupported
regions, and thus the facet list, has been filtered.  This is because
the adjacency information is computed from the total list of unsupported
facets, and because unsupported regions are mutually exclusive.