Rework project structure
Create a solution with multiple projects to share code between both a CLI and a
GUI implementation.  Do this with a common library.

Add the start of a GUI implementation that separates user interface definition
from logic.  Use Avalonia UI because this is cross-platform (with Linux
support) and is more mature than .NET MAUI.

Do not enable implicit usings.  Have everything be explicit.  While this does
increase the boilerplate, it makes dependencies in use more clear.