Get initial GUI together with controls and 3D rendering
Code was adapted from https://github.com/kekekeks/Avalonia-Silk.NET-Example on
the suggestion of
https://github.com/AvaloniaUI/Avalonia/discussions/6842#discussioncomment-6549959
to work with current Avalonia.

These changes set up the GUI so that user input controls are locked to the
right in a vertical fashion and the 3D rendering takes up the rest of the
space.